feat: draw paths between components
We use the paths calculated in the arrow service to draw paths
between dom elements on our SVG-"canvas". We need to make sure to
recalculate the paths when the container of the svg element changes
in size otherwise the paths will get out of sync with the positions
of bounding-boxes on the screen.